Add controller and remake render
